requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
django-leaflet
psycopg2-binary
gunicorn
whitenoise
celery
redis
django-celery-beat
twilio
sendgrid
stripe
django-import-export
django-crispy-forms
crispy-tailwind
pillow
django-cleanup
django-debug-toolbar
django-extensions
python-decouple
requests
httpx
openai
django-celery-results
channels
channels-redis
django-allauth
django-phonenumber-field
phonenumbers
daphne
drf-yasg
django-storages
boto3
certifi
dj-database-url
django-redis
hiredis
argon2-cffi
djangorestframework-simplejwt
django-sendgrid-v5
coreapi
google-generativeai
groq
openpyxl
reportlab
tenacity
apscheduler
